pyproject.toml
[project]
name = "lens-mcp"
version = "0.1.0"
description = "The MCP server for TTYF (Talk To Your Finances)"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.2.8",
    "keyring>=25.6.0",
    "plaid-python>=31.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
    "structlog>=25.3.0",
    "lens-cli==0.1.0",
]

[build-system]
requires = ["setuptools>=80.3.1"]
build-backend = "setuptools.build_meta"